Relocate invoice query tests to graphql crate
