+++ graphql/src/queries.rs
@@ -4,13 +4,13 @@ use graphql_core::standard_graphql_error::StandardGraphqlError;
 use graphql_core::ContextExt;
 use graphql_general_queries::store::*;
 use graphql_general_queries::*;
-use graphql_invoice::invoice_queries::*;
 use graphql_requisition::requisition_queries::*;
 use graphql_stocktake::stocktake_queries::*;
 use graphql_types::types::*;
 use repository::LocationFilter;
 use repository::PaginationOption;
 
+#[derive(Default)]
 pub struct Queries;
 
 #[Object]
@@ -118,37 +118,6 @@ impl Queries {
         sort: Option<Vec<ItemSortInput>>,
     ) -> Result<ItemsResponse> {
         items(ctx, page, filter, sort)
-    }
-
-    pub async fn invoice(
-        &self,
-        ctx: &Context<'_>,
-        store_id: String,
-        #[graphql(desc = "id of the invoice")] id: String,
-    ) -> Result<InvoiceResponse> {
-        get_invoice(ctx, Some(&store_id), &id)
-    }
-
-    pub async fn invoice_by_number(
-        &self,
-        ctx: &Context<'_>,
-        store_id: String,
-        invoice_number: u32,
-        r#type: InvoiceNodeType,
-    ) -> Result<InvoiceResponse> {
-        get_invoice_by_number(ctx, &store_id, invoice_number, r#type)
-    }
-
-    pub async fn invoices(
-        &self,
-        ctx: &Context<'_>,
-        store_id: String,
-        #[graphql(desc = "Pagination option (first and offset)")] page: Option<PaginationInput>,
-        #[graphql(desc = "Filter option")] filter: Option<InvoiceFilterInput>,
-        #[graphql(desc = "Sort options (only first sort input is evaluated for this endpoint)")]
-        sort: Option<Vec<InvoiceSortInput>>,
-    ) -> Result<InvoicesResponse> {
-        get_invoices(ctx, &store_id, page, filter, sort)
     }
 
     pub async fn invoice_counts(
